requirements.txt
requests==2.31.0
flask>=3.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
schedule==1.2.1
rich==13.7.1
